Fix clang-tidy warnings and refactor to fixed-size integers
- Resolved all clang-tidy warnings related to magic numbers, pointer arithmetic, reinterpret_cast, and narrowing conversions.
- Refactored project-owned code to use fixed-size integer types (int32_t, uint32_t, etc.) and size_t exclusively, as requested.
- Introduced named constants for magic numbers in date/time validation and UTF-8 checking.
- Applied clang-format to all modified files.
- Verified all tests pass.

+++ lib/certificate.cpp
@@ -81,8 +81,8 @@ void Certificate::printSubsectionHeader(const std::string& name)
     log("");
     log("┌────────────────────────────────────────┐");
     std::stringstream ss;
-    constexpr size_t SUBSECTION_NAME_WIDTH = 39;
-    ss << "│ " << std::left << std::setw(static_cast<int>(SUBSECTION_NAME_WIDTH)) << name << "│";
+    constexpr int32_t SUBSECTION_NAME_WIDTH = 39;
+    ss << "│ " << std::left << std::setw(static_cast<int32_t>(SUBSECTION_NAME_WIDTH)) << name << "│";
     log(ss.str());
     log("└────────────────────────────────────────┘");
     log("");
